add_user_id_to_chat_model

+++ api/sqlc_queries/chat_model.sql.go
@@ -10,7 +10,7 @@ import (
 )
 
 const chatModelByID = `-- name: ChatModelByID :one
-SELECT id, name, label, is_default, url, api_auth_header, api_auth_key FROM chat_model WHERE id = $1
+SELECT id, name, label, is_default, url, api_auth_header, api_auth_key, user_id FROM chat_model WHERE id = $1
 `
 
 func (q *Queries) ChatModelByID(ctx context.Context, id int32) (ChatModel, error) {
@@ -24,12 +24,13 @@ func (q *Queries) ChatModelByID(ctx context.Context, id int32) (ChatModel, error
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
 
 const chatModelByName = `-- name: ChatModelByName :one
-SELECT id, name, label, is_default, url, api_auth_header, api_auth_key FROM chat_model WHERE name = $1
+SELECT id, name, label, is_default, url, api_auth_header, api_auth_key, user_id FROM chat_model WHERE name = $1
 `
 
 func (q *Queries) ChatModelByName(ctx context.Context, name string) (ChatModel, error) {
@@ -43,14 +44,15 @@ func (q *Queries) ChatModelByName(ctx context.Context, name string) (ChatModel, 
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
 
 const createChatModel = `-- name: CreateChatModel :one
-INSERT INTO chat_model (name, label, is_default, url, api_auth_header, api_auth_key)
-VALUES ($1, $2, $3, $4, $5, $6)
-RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key
+INSERT INTO chat_model (name, label, is_default, url, api_auth_header, api_auth_key, user_id)
+VALUES ($1, $2, $3, $4, $5, $6, $7)
+RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key, user_id
 `
 
 type CreateChatModelParams struct {
@@ -60,6 +62,7 @@ type CreateChatModelParams struct {
 	Url           string
 	ApiAuthHeader string
 	ApiAuthKey    string
+	UserID        int32
 }
 
 func (q *Queries) CreateChatModel(ctx context.Context, arg CreateChatModelParams) (ChatModel, error) {
@@ -70,6 +73,7 @@ func (q *Queries) CreateChatModel(ctx context.Context, arg CreateChatModelParams
 		arg.Url,
 		arg.ApiAuthHeader,
 		arg.ApiAuthKey,
+		arg.UserID,
 	)
 	var i ChatModel
 	err := row.Scan(
@@ -80,21 +84,27 @@ func (q *Queries) CreateChatModel(ctx context.Context, arg CreateChatModelParams
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
 
 const deleteChatModel = `-- name: DeleteChatModel :exec
-DELETE FROM chat_model WHERE id = $1
+DELETE FROM chat_model WHERE id = $1 and user_id = $2
 `
 
-func (q *Queries) DeleteChatModel(ctx context.Context, id int32) error {
-	_, err := q.db.ExecContext(ctx, deleteChatModel, id)
+type DeleteChatModelParams struct {
+	ID     int32
+	UserID int32
+}
+
+func (q *Queries) DeleteChatModel(ctx context.Context, arg DeleteChatModelParams) error {
+	_, err := q.db.ExecContext(ctx, deleteChatModel, arg.ID, arg.UserID)
 	return err
 }
 
 const getDefaultChatModel = `-- name: GetDefaultChatModel :one
-SELECT id, name, label, is_default, url, api_auth_header, api_auth_key FROM chat_model WHERE is_default = true
+SELECT id, name, label, is_default, url, api_auth_header, api_auth_key, user_id FROM chat_model WHERE is_default = true
 `
 
 func (q *Queries) GetDefaultChatModel(ctx context.Context) (ChatModel, error) {
@@ -108,12 +118,13 @@ func (q *Queries) GetDefaultChatModel(ctx context.Context) (ChatModel, error) {
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
 
 const listChatModels = `-- name: ListChatModels :many
-SELECT id, name, label, is_default, url, api_auth_header, api_auth_key FROM chat_model ORDER BY id
+SELECT id, name, label, is_default, url, api_auth_header, api_auth_key, user_id FROM chat_model ORDER BY id
 `
 
 func (q *Queries) ListChatModels(ctx context.Context) ([]ChatModel, error) {
@@ -133,6 +144,7 @@ func (q *Queries) ListChatModels(ctx context.Context) ([]ChatModel, error) {
 			&i.Url,
 			&i.ApiAuthHeader,
 			&i.ApiAuthKey,
+			&i.UserID,
 		); err != nil {
 			return nil, err
 		}
@@ -149,8 +161,8 @@ func (q *Queries) ListChatModels(ctx context.Context) ([]ChatModel, error) {
 
 const updateChatModel = `-- name: UpdateChatModel :one
 UPDATE chat_model SET name = $2, label = $3, is_default = $4, url = $5, api_auth_header = $6, api_auth_key = $7
-WHERE id = $1
-RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key
+WHERE id = $1 and user_id = $8
+RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key, user_id
 `
 
 type UpdateChatModelParams struct {
@@ -161,6 +173,7 @@ type UpdateChatModelParams struct {
 	Url           string
 	ApiAuthHeader string
 	ApiAuthKey    string
+	UserID        int32
 }
 
 func (q *Queries) UpdateChatModel(ctx context.Context, arg UpdateChatModelParams) (ChatModel, error) {
@@ -172,6 +185,7 @@ func (q *Queries) UpdateChatModel(ctx context.Context, arg UpdateChatModelParams
 		arg.Url,
 		arg.ApiAuthHeader,
 		arg.ApiAuthKey,
+		arg.UserID,
 	)
 	var i ChatModel
 	err := row.Scan(
@@ -182,6 +196,7 @@ func (q *Queries) UpdateChatModel(ctx context.Context, arg UpdateChatModelParams
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
@@ -189,7 +204,7 @@ func (q *Queries) UpdateChatModel(ctx context.Context, arg UpdateChatModelParams
 const updateChatModelKey = `-- name: UpdateChatModelKey :one
 UPDATE chat_model SET api_auth_key = $2
 WHERE id = $1
-RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key
+RETURNING id, name, label, is_default, url, api_auth_header, api_auth_key, user_id
 `
 
 type UpdateChatModelKeyParams struct {
@@ -208,6 +223,7 @@ func (q *Queries) UpdateChatModelKey(ctx context.Context, arg UpdateChatModelKey
 		&i.Url,
 		&i.ApiAuthHeader,
 		&i.ApiAuthKey,
+		&i.UserID,
 	)
 	return i, err
 }
